Extract youtube player to separate file
